+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Persistence/Dbal/Query/DbalExtractConnectionsProductEventCountQuery.php
@@ -137,7 +137,6 @@ FROM (
 INNER JOIN oro_user u ON u.username = author AND u.user_type = :user_type
 INNER JOIN akeneo_connectivity_connection conn ON conn.user_id = u.id
 WHERE conn.auditable = 1 AND conn.flow_type = :flow_type
-GROUP BY conn.code;
 SQL;
         $dataRows = $this->dbalConnection->executeQuery(
             $sqlQuery,