+++ src/Akeneo/Pim/Enrichment/Bundle/EventSubscriber/EntityWithAssociations/PersistTwoWayAssociationSubscriber.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Pim\Enrichment\Bundle\EventSubscriber\EntityWithAssociations;
+
+use Akeneo\Pim\Enrichment\Component\Product\Model\AssociationInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithAssociationsInterface;
+use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
+use Akeneo\Tool\Component\StorageUtils\StorageEvents;
+use Doctrine\Common\Persistence\ManagerRegistry;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\EventDispatcher\GenericEvent;
+
+final class PersistTwoWayAssociationSubscriber implements EventSubscriberInterface
+{
+    /** @var ManagerRegistry */
+    private $registry;
+
+    public function __construct(
+        ManagerRegistry $registry
+    ) {
+        $this->registry = $registry;
+    }
+
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            StorageEvents::PRE_SAVE => 'handlePreSave',
+        ];
+    }
+
+    public function handlePreSave(GenericEvent $event): void
+    {
+        $entity = $event->getSubject();
+
+        if (!$entity instanceof EntityWithAssociationsInterface) {
+            return;
+        }
+
+        /** @var AssociationInterface $association */
+        foreach ($entity->getAssociations() as $association) {
+            $associationType = $association->getAssociationType();
+
+            if (!$associationType->isTwoWay()) {
+                continue;
+            }
+
+            foreach ($association->getProducts() as $product) {
+                $this->persistInversedAssociation($associationType, $product);
+            }
+
+            foreach ($association->getProductModels() as $productModel) {
+                $this->persistInversedAssociation($associationType, $productModel);
+            }
+        }
+    }
+
+    private function persistInversedAssociation(
+        AssociationTypeInterface $associationType,
+        EntityWithAssociationsInterface $associatedEntity
+    ): void {
+        $em = $this->registry->getManager();
+
+        $inversedAssociation = $associatedEntity->getAssociationForType($associationType);
+
+        if (null !== $inversedAssociation) {
+            $em->persist($inversedAssociation);
+        }
+    }
+}