+++ tests/back/Pim/Structure/EndToEnd/AssociationType/ExternalApi/PartialUpdateListAssociationTypeEndToEnd.php
@@ -36,10 +36,12 @@ JSON;
                     'en_US' => 'Cross sell',
                     'fr_FR' => 'Vente croisée',
                 ],
+                'is_two_way' => false,
             ],
             'NEW_SELL' => [
                 'code'   => 'NEW_SELL',
                 'labels' => [],
+                'is_two_way' => false,
             ],
         ];